+++ src/views/Home.vue
@@ -2,7 +2,7 @@
   <h1 class="mb-1">{{ $t('Home.Welcome.Title') }}</h1>
   <p>{{ $t('Home.Welcome.Subtitle') }}</p>
   <br />
-  <v-btn prepend-icon="mdi-plus" to="/add">{{ $t('Home.AddPrice') }}</v-btn>
+  <v-btn color="primary" prepend-icon="mdi-plus" to="/add">{{ $t('Home.AddPrice') }}</v-btn>
 
   <v-snackbar
     v-model="settingsSuccessMessage"