+++ src/views/ProductDetail.vue
@@ -7,7 +7,8 @@
   </v-row>
 
   <v-row class="mt-0" v-if="!productNotFound">
-    <v-col cols="12" sm="6">
+    <v-col cols="12">
+      <v-btn class="mr-2" size="small" color="primary" prepend-icon="mdi-plus" to="/add">Add a price</v-btn>
       <v-btn v-if="product.code && product.source" size="small" append-icon="mdi-open-in-new" :href="getProductOFFUrl(product)" target="_blank">
         Open Food Facts
       </v-btn>
@@ -40,7 +41,7 @@
 
   <v-row v-if="productPriceList.length < productPriceTotal" class="mb-2">
     <v-col align="center">
-      <v-btn size="small" @click="getProductPrices">{{ $t('ProductDetail.LoadMore') }}</v-btn>
+      <v-btn size="small" :loading="loading @click="getProductPrices">{{ $t('ProductDetail.LoadMore') }}</v-btn>
     </v-col>
   </v-row>
 </template>